fix(cellNav): Don't lose focus with wrapping nav
When navigating beyond the left/right edges of grid around to the other
side, the grid was losing focus which prevented further navigation because
any keydown events were being sent to <body> (the new activeElement). This
fix causes the renderContainer to be REFOCUSED after a GRID_SCROLL event
in the case that the activeElement is document.body. We are assuming that
when scrolling within the grid, an element within the grid should have
focus.